Code cleaning + improvement
