+++ backend/dissemination/summary_reports.py
@@ -1,9 +1,9 @@
 from datetime import datetime
-import openpyxl as pyxl
 import io
 import logging
 import uuid
 import time
+import openpyxl as pyxl
 
 from boto3 import client as boto3_client
 from botocore.client import ClientError, Config
@@ -25,6 +25,7 @@ from dissemination.models import (
     Note,
     Passthrough,
     SecondaryAuditor,
+    DisseminationCombined,
 )
 
 logger = logging.getLogger(__name__)
@@ -178,7 +179,7 @@ field_name_ordered = {
         "award_reference",
         "federal_agency_prefix",
         "federal_award_extension",
-        "_aln",
+        "aln",
         "findings_count",
         "additional_award_identification",
         "federal_program_name",
@@ -198,9 +199,9 @@ field_name_ordered = {
     ],
     "finding": [
         "report_id",
-        "_federal_agency_prefix",
-        "_federal_award_extension",
-        "_aln",
+        "federal_agency_prefix",
+        "federal_award_extension",
+        "aln",
         "award_reference",
         "reference_number",
         "type_requirement",
@@ -266,7 +267,14 @@ can_read_tribal_disclaimer = "This document includes one or more Tribal entities
 cannot_read_tribal_disclaimer = "This document includes one or more Tribal entities that have chosen to keep their data private per 2 CFR 200.512(b)(2). It doesn't include their audit findings text, corrective action plan, or notes to SEFA."
 
 
-def _get_tribal_report_ids(report_ids):
+def _get_model_by_name(name):
+    for m in models:
+        if m.__name__.lower() == name:
+            return m
+    return None
+
+
+def get_tribal_report_ids(report_ids):
     t0 = time.time()
     """Filters the given report_ids with only ones that are tribal"""
     objects = General.objects.all().filter(report_id__in=report_ids, is_public=False)
@@ -335,37 +343,10 @@ def insert_dissem_coversheet(workbook, contains_tribal, include_private):
 
 
 def _get_attribute_or_data(obj, field_name):
-    if field_name.startswith("_"):
-        if field_name == "_aln":
-            if isinstance(obj, FederalAward):
-                return (
-                    getattr(obj, "federal_agency_prefix")
-                    + "."
-                    + getattr(obj, "federal_award_extension")
-                )
-            elif isinstance(obj, Finding):
-                fa = FederalAward.objects.get(
-                    report_id=getattr(obj, "report_id"),
-                    award_reference=getattr(obj, "award_reference"),
-                )
-                return (
-                    getattr(fa, "federal_agency_prefix")
-                    + "."
-                    + getattr(fa, "federal_award_extension")
-                )
-        else:
-            field_name = field_name[1:]
-            if isinstance(obj, Finding):
-                fa = FederalAward.objects.get(
-                    report_id=getattr(obj, "report_id"),
-                    award_reference=getattr(obj, "award_reference"),
-                )
-                return getattr(fa, field_name)
-    else:
-        value = getattr(obj, field_name)
-        if isinstance(value, General):
-            value = value.report_id
-        return value
+    value = getattr(obj, field_name)
+    if isinstance(value, General):
+        value = value.report_id
+    return value
 
 
 def gather_report_data_dissemination(report_ids, tribal_report_ids, include_private):
@@ -385,40 +366,128 @@ def gather_report_data_dissemination(report_ids, tribal_report_ids, include_priv
     }
     """
     t0 = time.time()
-
     # Make report IDs unique
     report_ids = set(report_ids)
+    all_names = set(field_name_ordered.keys())
+    names_in_dc = set(["general", "federalaward", "finding", "passthrough"])
+    names_not_in_dc = all_names - names_in_dc
+    data = initialize_data_structure(names_in_dc.union(names_not_in_dc))
 
+    process_combined_results(
+        report_ids, names_in_dc, data, include_private, tribal_report_ids
+    )
+
+    process_non_combined_results(
+        report_ids, names_not_in_dc, data, include_private, tribal_report_ids
+    )
+
+    return (data, time.time() - t0)
+
+
+def initialize_data_structure(names):
     data = {}
+    for model_name in names:
+        data[model_name] = {
+            "field_names": field_name_ordered[model_name],
+            "entries": [],
+        }
+    return data
 
-    for model in models:
-        model_name = model.__name__.lower()
 
-        # This pulls directly from the model
-        #   fields = model._meta.get_fields()
-        #   field_names = [f.name for f in fields]
-        # This uses the ordered columns above
+def process_combined_results(
+    report_ids, names_in_dc, data, include_private, tribal_report_ids
+):
+    # Grab all the rows from the combined table into a local structure.
+    # We'll do this in memory. This table flattens general, federalaward, and findings
+    # so we can move much faster on those tables without extra lookups.
+    dc_results = DisseminationCombined.objects.all().filter(report_id__in=report_ids)
+    # Different tables want to be visited/filtered differently.
+    visited = set()
+    # Do all of the names in the DisseminationCombined at the same time.
+    # That way, we only go through the results once.
+    for obj in dc_results:
+        for model_name in names_in_dc:
+            field_names = field_name_ordered[model_name]
+            report_id = getattr(obj, "report_id")
+            award_reference = getattr(obj, "award_reference")
+            reference_number = getattr(obj, "reference_number")
+            passthrough_name = getattr(obj, "passthrough_name")
+
+            # WATCH THIS IF/ELIF
+            # It is making sure we do not double-disseminate some rows.
+            ####
+            # GENERAL
+            if model_name == "general" and report_id in visited:
+                pass
+            ####
+            # PASSTHROUGH
+            # We should never disseminate something that has no name.
+            elif model_name == "passthrough" and passthrough_name is None:
+                pass
+            ####
+            # FEDERAL AWARD
+            # This condition is actually filtering out the damage to the
+            # data from the race hazard we had at the start of 2024.
+            # NOTE
+            # We cannot filter `passthrough` here. Each award reference row has
+            # a one-to-many relationship with passthrough.
+            elif (
+                model_name == "federalaward"
+                and f"{report_id}-{award_reference}" in visited
+            ):
+                pass
+            ####
+            # FINDING
+            elif model_name == "finding" and (
+                award_reference is None or reference_number is None
+            ):
+                # And we don't include rows in finding where there are none.
+                pass
+            else:
+                # Track to limit duplication
+                if model_name == "general":
+                    visited.add(report_id)
+                # Handle special tracking for federal awards, so we don't duplicate award # rows.
+                if model_name == "federalaward":
+                    visited.add(f"{report_id}-{award_reference}")
+                # Omit rows for private tribal data when the user doesn't have perms
+                if (
+                    model_name in restricted_model_names
+                    and not include_private
+                    and report_id in tribal_report_ids
+                ):
+                    pass
+                else:
+                    data[model_name]["entries"].append(
+                        [getattr(obj, field_name) for field_name in field_names]
+                    )
+
+
+def process_non_combined_results(
+    report_ids, names_not_in_dc, data, include_private, tribal_report_ids
+):
+    for model_name in names_not_in_dc:
+        model = _get_model_by_name(model_name)
+        print(model_name)
         field_names = field_name_ordered[model_name]
-
-        data[model_name] = {"field_names": field_names, "entries": []}
-
         objects = model.objects.all().filter(report_id__in=report_ids)
+        # Walk the objects
         for obj in objects:
             report_id = _get_attribute_or_data(obj, "report_id")
-
             # Omit rows for private tribal data when the user doesn't have perms
             if (
                 model_name in restricted_model_names
                 and not include_private
                 and report_id in tribal_report_ids
             ):
-                continue
-
-            data[model_name]["entries"].append(
-                [_get_attribute_or_data(obj, field_name) for field_name in field_names]
-            )
-    t1 = time.time()
-    return (data, t1 - t0)
+                pass
+            else:
+                data[model_name]["entries"].append(
+                    [
+                        _get_attribute_or_data(obj, field_name)
+                        for field_name in field_names
+                    ]
+                )
 
 
 def gather_report_data_pre_certification(i2d_data):
@@ -564,7 +633,7 @@ def persist_workbook(workbook):
 
 def generate_summary_report(report_ids, include_private=False):
     t0 = time.time()
-    (tribal_report_ids, ttri) = _get_tribal_report_ids(report_ids)
+    (tribal_report_ids, ttri) = get_tribal_report_ids(report_ids)
     (data, tgrdd) = gather_report_data_dissemination(
         report_ids, tribal_report_ids, include_private
     )
